--- 2l_SD/2l_SD/FriendsCell.swift
@@ -1,15 +1,0 @@
-//
-//  FriendsCell.swift
-//  2l_SD
-//
-//  Created by Irina Semenova on 01/12/2019.
-//  Copyright © 2019 Denis Semenov. All rights reserved.
-//
-
-import UIKit
-
-class FriendsCell: UITableViewCell {
-    
-    @IBOutlet weak var friendName: UILabel!
-    
-}
