+++ SwiftUIApp/SwiftUIApp.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "self:/Users/Si/Documents/Apps/GeekBrains/cource/swift/webinar_02/lesson_02/2l_SD/SwiftUIApp.xcodeproj">
+   </FileRef>
+</Workspace>
