+++ SwiftUIApp/SwiftUIApp.xcodeproj/xcuserdata/Si.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   type = "1"
+   version = "2.0">
+   <Breakpoints>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            shouldBeEnabled = "No"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "2l_SD/ViewController.swift"
+            timestampString = "597022149.889675"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "31"
+            endingLineNumber = "31"
+            landmarkName = "ViewController"
+            landmarkType = "3">
+            <Locations>
+               <Location
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "_l_SD.ViewController.userLogin.getter : Swift.String"
+                  moduleName = "2l_SD"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/Si/Documents/Apps/GeekBrains/cource/swift/webinar_02/lesson_02/2l_SD/2l_SD/ViewController.swift"
+                  timestampString = "597017412.862726"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "31"
+                  endingLineNumber = "31"
+                  offsetFromSymbolStart = "53">
+               </Location>
+               <Location
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "_l_SD.ViewController.init(nibName: Swift.Optional&lt;Swift.String&gt;, bundle: Swift.Optional&lt;__C.NSBundle&gt;) -&gt; _l_SD.ViewController"
+                  moduleName = "2l_SD"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/Si/Documents/Apps/GeekBrains/cource/swift/webinar_02/lesson_02/2l_SD/2l_SD/ViewController.swift"
+                  timestampString = "597017412.8667949"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "31"
+                  endingLineNumber = "31"
+                  offsetFromSymbolStart = "399">
+               </Location>
+               <Location
+                  shouldBeEnabled = "Yes"
+                  ignoreCount = "0"
+                  continueAfterRunningActions = "No"
+                  symbolName = "_l_SD.ViewController.init(coder: __C.NSCoder) -&gt; Swift.Optional&lt;_l_SD.ViewController&gt;"
+                  moduleName = "2l_SD"
+                  usesParentBreakpointCondition = "Yes"
+                  urlString = "file:///Users/Si/Documents/Apps/GeekBrains/cource/swift/webinar_02/lesson_02/2l_SD/2l_SD/ViewController.swift"
+                  timestampString = "597017412.869101"
+                  startingColumnNumber = "9223372036854775807"
+                  endingColumnNumber = "9223372036854775807"
+                  startingLineNumber = "31"
+                  endingLineNumber = "31"
+                  offsetFromSymbolStart = "298">
+               </Location>
+            </Locations>
+         </BreakpointContent>
+      </BreakpointProxy>
+   </Breakpoints>
+</Bucket>
