+++ SwiftUIApp/SwiftUIApp/Assets.xcassets/AppIcon.appiconset/Contents.json
