+++ SwiftUIApp/SwiftUIApp/Assets.xcassets/Contents.json
