+++ SwiftUIApp/SwiftUIApp/Assets.xcassets/copy2.imageset/Contents.json
