+++ SwiftUIApp/SwiftUIApp/Assets.xcassets/copy2.imageset/copy2.JPG
