+++ SwiftUIApp/SwiftUIApp/Assets.xcassets/info.imageset/Contents.json
