+++ SwiftUIApp/SwiftUIApp/Assets.xcassets/msg.imageset/Contents.json
